monolithic app
will add features then split code
