+++ README.md
@@ -13,18 +13,26 @@ Justification:
 I'm very pleased with the development workflow. The problem is when it comes time to serve the static site via GitHub Pages, for security reasons (cross-origin stuff being disabled by GitHub and therefore out of my power to control). From the current repo, I have no problem connecting the index.html file to the asa55.github.io/market-segmentation-tool URI. The problem is that all other relative references (every other static file referenced by this index.html) lives in a place that by default the Gatsby build process doesn't expect. But even if I point to the files in this repo, I get some CORS blocking errors. If I were doing this in my asa55.github.io repo, everything would be dandy. But I don't want to do that. Why? Because asa55.github.io is its own site, with its own code. I want to maintain that separately as much as humanly possible. But I also want to leverage the fact that I have free hosting through GitHub pages. The solution I landed on was to generate my static files in this repo, then manually move them over to the asa55.github.io repo. 
 
 Right after you first build the container:
->> cd gatsby-site
->> npm install
+
+```>> cd gatsby-site```
+
+```>> npm install```
 
 To develop:
->> cd gatsby-site
->> gatsby develop
+
+```>> cd gatsby-site```
+
+```>> gatsby develop```
     (dev server runs on localhost:8000)
 
 To deploy:
->> npm run deploy
+
+```>> npm run deploy```
+
 Copy the files from [this-repo]/gatsby-site/public/ into [asa55.github.io]/market-segmentation-tool
 For more info on what I'm doing here: https://www.gatsbyjs.com/docs/how-gatsby-works-with-github-pages/
 
 
 Note to self: recharts came with the dashboard template I pulled from the official material-ui site. If I decide not to use it, remember to ```npm uninstall recharts```
+
+Another note to self: building changes takes forever. I need a faster way to see my results. React is component oriented, so I'd be okay developing snippets elsewhere then figuring out how to integrate them into my site. People seem to like codesandbox.io. There are probably other local dev options.
