+++ gatsby-site/package.json
@@ -21,6 +21,7 @@
     "gatsby-theme-material-ui": "^1.0.13",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
+    "react-redux": "^7.2.2",
     "recharts": "^1.8.5"
   },
   "devDependencies": {
